requirements.txt
Flask
mysql-connector-python
pandas
matplotlib
python-dotenv